% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addpercent.R
\name{addPercent}
\alias{addPercent}
\title{Format a number to percentage with required decimal places}
\usage{
addPercent(x, d)
}
\arguments{
\item{x}{The number to be formatted,}

\item{d}{The number of decimal places that should be retained after formatting. If parameter d is not provided, then the default is zero.}
}
\value{
A string that is \code{x} with percentage sign, %.
}
\description{
This function accepts up to two parameters and returns a string with a percentage sign attached.
}
\examples{
\dontrun{
addPercent(0.65423, 1)
addPercent(0.65423)
sapply(c(123.4546, .876589, 0.45678),addPercent, d=2)
}
}
