% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsig.R
\name{signifDigit}
\alias{signifDigit}
\title{Round and keep the significant digits to the right of a decimal}
\usage{
signifDigit(x, d)
}
\arguments{
\item{x}{The number to be formatted,}

\item{d}{The number of significant figures that should be retained after formatting. If parameter d is not provided, then the function will fail.}
}
\value{
A string that is \code{x}.
}
\description{
This function allows the user to specify a number of significant digits, keeps significant zeroes
and the digits to the left of the decimal point.
}
\details{
This function relies on the stringr library. More specifically, it utilizes the str_pad function within stringr to pad with zeros when there are decimals present after rounding.
The function will also use grepl to test whether there are decimals within the formatted number and if not will return all the same digits but as a string.
}
\examples{
\dontrun{
signifDigit(5423.63875,2)
sapply(c(123.4546, 87.6589, 9.5678),signifDigit,d=3)
}
}
\seealso{
\code{\link{grepl}}

\code{\link{prettyNum}}
}
